src/test/java/lt/justas/demo/app/controller/ApiControllerTest.java
package lt.justas.demo.app.controller;


import lt.justas.demo.app.service.ApiService;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.web.servlet.MockMvc;

import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(controllers = ApiController.class)
public class ApiControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private ApiService apiService;

    @Test
    public void apiSearchEndpointIsExposed() throws Exception {
        mockMvc.perform(get("/api/search")
                .param("artistName", "abba")
        ).andExpect(status().isOk());
        verify(apiService).searchArtistsByName("abba");
    }
}
